Refactor: move FileSettingsHealthInfo and FileSettingsHealthTracker to top-level classes
