+++ server/src/main/java/org/elasticsearch/health/node/HealthInfoCache.java
@@ -17,12 +17,11 @@ import org.elasticsearch.cluster.node.DiscoveryNode;
 import org.elasticsearch.cluster.service.ClusterService;
 import org.elasticsearch.core.Nullable;
 import org.elasticsearch.health.node.selection.HealthNode;
-import org.elasticsearch.reservedstate.service.FileSettingsService;
 
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
-import static org.elasticsearch.reservedstate.service.FileSettingsService.FileSettingsHealthInfo.INDETERMINATE;
+import static org.elasticsearch.health.node.FileSettingsHealthInfo.INDETERMINATE;
 
 /**
  * Keeps track of several health statuses per node that can be used in health.
@@ -34,7 +33,7 @@ public class HealthInfoCache implements ClusterStateListener {
     @Nullable
     private volatile DataStreamLifecycleHealthInfo dslHealthInfo = null;
     private volatile ConcurrentHashMap<String, RepositoriesHealthInfo> repositoriesInfoByNode = new ConcurrentHashMap<>();
-    private volatile FileSettingsService.FileSettingsHealthInfo fileSettingsHealthInfo = INDETERMINATE;
+    private volatile FileSettingsHealthInfo fileSettingsHealthInfo = INDETERMINATE;
 
     private HealthInfoCache() {}
 
@@ -49,7 +48,7 @@ public class HealthInfoCache implements ClusterStateListener {
         @Nullable DiskHealthInfo diskHealthInfo,
         @Nullable DataStreamLifecycleHealthInfo latestDslHealthInfo,
         @Nullable RepositoriesHealthInfo repositoriesHealthInfo,
-        @Nullable FileSettingsService.FileSettingsHealthInfo fileSettingsHealthInfo
+        @Nullable FileSettingsHealthInfo fileSettingsHealthInfo
     ) {
         if (diskHealthInfo != null) {
             diskInfoByNode.put(nodeId, diskHealthInfo);
