+++ server/src/test/java/org/elasticsearch/health/node/DiskHealthIndicatorServiceTests.java
@@ -36,7 +36,6 @@ import org.elasticsearch.health.HealthStatus;
 import org.elasticsearch.health.ImpactArea;
 import org.elasticsearch.index.IndexVersion;
 import org.elasticsearch.index.shard.ShardId;
-import org.elasticsearch.reservedstate.service.FileSettingsService;
 import org.elasticsearch.test.ESTestCase;
 import org.elasticsearch.xcontent.ToXContent;
 import org.elasticsearch.xcontent.XContentBuilder;
@@ -279,7 +278,7 @@ public class DiskHealthIndicatorServiceTests extends ESTestCase {
             diskInfoByNode,
             DataStreamLifecycleHealthInfo.NO_DSL_ERRORS,
             Map.of(),
-            FileSettingsService.FileSettingsHealthInfo.INDETERMINATE
+            FileSettingsHealthInfo.INDETERMINATE
         );
 
         HealthIndicatorResult result = diskHealthIndicatorService.calculate(true, healthInfo);
@@ -1056,12 +1055,7 @@ public class DiskHealthIndicatorServiceTests extends ESTestCase {
                 diskInfoByNode.put(node.getId(), diskHealthInfo);
             }
         }
-        return new HealthInfo(
-            diskInfoByNode,
-            DataStreamLifecycleHealthInfo.NO_DSL_ERRORS,
-            Map.of(),
-            FileSettingsService.FileSettingsHealthInfo.INDETERMINATE
-        );
+        return new HealthInfo(diskInfoByNode, DataStreamLifecycleHealthInfo.NO_DSL_ERRORS, Map.of(), FileSettingsHealthInfo.INDETERMINATE);
     }
 
     private static ClusterService createClusterService(Collection<DiscoveryNode> nodes, boolean withBlockedIndex) {
