+++ server/src/test/java/org/elasticsearch/health/node/FetchHealthInfoCacheActionTests.java
@@ -18,7 +18,6 @@ import org.elasticsearch.cluster.node.DiscoveryNodeRole;
 import org.elasticsearch.cluster.node.DiscoveryNodeUtils;
 import org.elasticsearch.cluster.service.ClusterService;
 import org.elasticsearch.health.HealthStatus;
-import org.elasticsearch.reservedstate.service.FileSettingsService;
 import org.elasticsearch.test.ESTestCase;
 import org.elasticsearch.test.EqualsHashCodeTestUtils;
 import org.elasticsearch.test.transport.CapturingTransport;
@@ -132,7 +131,7 @@ public class FetchHealthInfoCacheActionTests extends ESTestCase {
                 new DiskHealthInfo(randomFrom(HealthStatus.values()), randomFrom(DiskHealthInfo.Cause.values())),
                 randomDslHealthInfo(),
                 randomRepoHealthInfo(),
-                FileSettingsService.FileSettingsHealthInfo.INDETERMINATE
+                FileSettingsHealthInfo.INDETERMINATE
             );
         }
         return healthInfoCache;
